Refactor some internal task-collection functions
